# docs/openapi-specification.yaml
openapi: 3.0.3
info:
  title: Thai Auto Insurance API
  version: 1.0.0
  description: |
    Comprehensive REST API for Thai Auto Insurance application supporting policy management, claims processing, and customer service operations.
    
    ## Features
    - Multi-language support (Thai/English)
    - JWT-based authentication
    - Thai-specific validations (National ID, phone numbers, addresses)
    - Real-time payment processing
    - Comprehensive claims management
    - Regulatory compliance (OIC requirements)
    
  contact:
    name: API Support Team
    email: api-support@autoinsurance.co.th
    url: https://docs.autoinsurance.co.th
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  termsOfService: https://autoinsurance.co.th/terms
servers:
  - url: https://api.autoinsurance.co.th/api/v1
    description: Production server
  - url: https://api-staging.autoinsurance.co.th/api/v1
    description: Staging server
  - url: https://api-dev.autoinsurance.co.th/api/v1
    description: Development server
security:
  - BearerAuth: []
...
paths:
...
  /admin/customers:
    get:
      tags: [Admin Operations]
      summary: Get all customers (Admin)
      description: Retrieve paginated list of all customers
      operationId: getAllCustomers
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 50
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [ACTIVE, INACTIVE, SUSPENDED]
      responses:
        '200':
          description: Customers retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminCustomerListResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
...
    AdminCustomerListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            customers:
              type: array
              items:
                $ref: '#/components/schemas/AdminCustomerSummary'
            pagination:
              $ref: '#/components/schemas/Pagination'
        timestamp:
          type: string
          format: date-time
    AdminCustomerSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
        phoneNumber:
          type: string
        status:
          type: string
        kycStatus:
          type: string
        totalPolicies:
          type: integer
        totalClaims:
          type: integer
        createdAt:
          type: string
          format: date-time
        lastLoginAt:
          type: string
          format: date-time
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean
...